Refactor table display logic
Move `displayRulesTable`, `formatAction`, and `formatConditionGroups` to a new UI module for better separation of concerns. Update imports in command files to use the new module. This improves code organization and reusability by centralizing table formatting logic. Also, fix import paths for `logger` to ensure consistent logging.
