+++ package.json
@@ -49,8 +49,8 @@
     "@commitlint/config-conventional": "^18.6.3",
     "@jest/globals": "^29.7.0",
     "@tsconfig/node20": "^20.1.4",
-    "@types/jest": "^29.5.12",
     "@types/find-up": "^4.0.2",
+    "@types/jest": "^29.5.12",
     "@types/node": "^20.12.12",
     "@types/prompts": "^2.4.9",
     "@types/signale": "^1.4.7",
@@ -70,8 +70,8 @@
     "rimraf": "^5.0.7",
     "semantic-release": "^23.1.1",
     "ts-jest": "^29.1.4",
-    "ts-node": "^10.9.2",
     "ts-json-schema-generator": "^2.3.0",
+    "ts-node": "^10.9.2",
     "tsup": "^8.0.2",
     "typescript": "^5.4.5"
   },
